# api-spec.yaml
...
openapi: "3.0.1"
info: 
  title: "Your API"
  version: "v1"
paths: 
...
  /api/Review: 
    get: 
      tags: 
      - "Review"
      responses: 
        200: 
          description: "OK"
          content: 
            text/plain: 
              schema: 
                type: "array"
                items: 
                  $ref: "#/components/schemas/Review"
            application/json: 
              schema: 
                type: "array"
                items: 
                  $ref: "#/components/schemas/Review"
            text/json: 
              schema: 
                type: "array"
                items: 
                  $ref: "#/components/schemas/Review"
        404: 
          description: "Not Found"
    post: 
      tags: 
      - "Review"
      requestBody: 
        content: 
          application/json: 
            schema: 
              $ref: "#/components/schemas/ReviewDTO"
          text/json: 
            schema: 
              $ref: "#/components/schemas/ReviewDTO"
          application/*+json: 
            schema: 
              $ref: "#/components/schemas/ReviewDTO"
      responses: 
        201: 
          description: "Created"
          content: 
            text/plain: 
              schema: 
                $ref: "#/components/schemas/Review"
            application/json: 
              schema: 
                $ref: "#/components/schemas/Review"
            text/json: 
              schema: 
                $ref: "#/components/schemas/Review"
        400: 
          description: "Bad Request"
        401: 
          description: "Unauthorized"
        404: 
          description: "Not Found"
  /api/Review/{id}: 
    get: 
      tags: 
      - "Review"
      parameters: 
      - name: "id"
        in: "path"
        required: "true"
        schema: 
          type: "integer"
          format: "int32"
      responses: 
        200: 
          description: "OK"
          content: 
            text/plain: 
              schema: 
                $ref: "#/components/schemas/Review"
            application/json: 
              schema: 
                $ref: "#/components/schemas/Review"
            text/json: 
              schema: 
                $ref: "#/components/schemas/Review"
        404: 
          description: "Not Found"
    put: 
      tags: 
      - "Review"
      parameters: 
      - name: "id"
        in: "path"
        required: "true"
        schema: 
          type: "integer"
          format: "int32"
      requestBody: 
        content: 
          application/json: 
            schema: 
              $ref: "#/components/schemas/ReviewDTO_Put"
          text/json: 
            schema: 
              $ref: "#/components/schemas/ReviewDTO_Put"
          application/*+json: 
            schema: 
              $ref: "#/components/schemas/ReviewDTO_Put"
      responses: 
        200: 
          description: "OK"
          content: 
            text/plain: 
              schema: 
                $ref: "#/components/schemas/Review"
            application/json: 
              schema: 
                $ref: "#/components/schemas/Review"
            text/json: 
              schema: 
                $ref: "#/components/schemas/Review"
        400: 
          description: "Bad Request"
        401: 
          description: "Unauthorized"
        403: 
          description: "Forbidden"
        404: 
          description: "Not Found"
    delete: 
      tags: 
      - "Review"
      parameters: 
      - name: "id"
        in: "path"
        required: "true"
        schema: 
          type: "integer"
          format: "int32"
      responses: 
        200: 
          description: "OK"
          content: 
            text/plain: 
              schema: 
                $ref: "#/components/schemas/Review"
            application/json: 
              schema: 
                $ref: "#/components/schemas/Review"
            text/json: 
              schema: 
                $ref: "#/components/schemas/Review"
        401: 
          description: "Unauthorized"
        403: 
          description: "Forbidden"
        404: 
          description: "Not Found"
components: 
  schemas: 
...
    Review: 
      required: 
      - "userId"
      type: "object"
      properties: 
        id: 
          type: "integer"
          format: "int32"
        rating: 
          maximum: "5"
          minimum: "1"
          type: "integer"
          format: "int32"
        text: 
          type: "string"
          nullable: "true"
        isPrivate: 
          type: "boolean"
        userId: 
          minLength: "1"
          type: "string"
        user: 
          $ref: "#/components/schemas/User"
        locationID: 
          type: "integer"
          format: "int32"
      additionalProperties: "false"
    ReviewDTO: 
      type: "object"
      properties: 
        rating: 
          maximum: "5"
          minimum: "1"
          type: "integer"
          format: "int32"
        text: 
          type: "string"
          nullable: "true"
        isPrivate: 
          type: "boolean"
        locationID: 
          type: "integer"
          format: "int32"
      additionalProperties: "false"
    ReviewDTO_Put: 
      type: "object"
      properties: 
        rating: 
          maximum: "5"
          minimum: "1"
          type: "integer"
          format: "int32"
        text: 
          type: "string"
          nullable: "true"
        isPrivate: 
          type: "boolean"
      additionalProperties: "false"
...
    User: 
      type: "object"
      properties: 
        id: 
          type: "string"
          nullable: "true"
        userName: 
          type: "string"
          nullable: "true"
        normalizedUserName: 
          type: "string"
          nullable: "true"
        email: 
          type: "string"
          nullable: "true"
        normalizedEmail: 
          type: "string"
          nullable: "true"
        emailConfirmed: 
          type: "boolean"
        passwordHash: 
          type: "string"
          nullable: "true"
        securityStamp: 
          type: "string"
          nullable: "true"
        concurrencyStamp: 
          type: "string"
          nullable: "true"
        phoneNumber: 
          type: "string"
          nullable: "true"
        phoneNumberConfirmed: 
          type: "boolean"
        twoFactorEnabled: 
          type: "boolean"
        lockoutEnd: 
          type: "string"
          format: "date-time"
          nullable: "true"
        lockoutEnabled: 
          type: "boolean"
        accessFailedCount: 
          type: "integer"
          format: "int32"
      additionalProperties: "false"